align everything use from_spark_recoverable
